# Speed up named type lookups in managed type system