@jkotas 2023-04-04
Also, would it be better to use the split IBC data per assembly, like what we do for R2R?

Reading raw MIBC files make sense when they closely match the app. It makes less sense when the MIBC files represent some generic profile and 90%+ of the data is irrelevant to the app being compiled.