@MichalStrehovsky 2023-04-05
> Would it be simpler and faster to just materialize all strings as `string` and use regular `Dictionary` for the lookup? It uses a bit more memory, but not that much more memory.

It would be simpler, but it doesn't look faster. After adding antivirus exclusion for ilc.exe and the folders around, and carefully listening to when CPU fans to avoid thermal throttling, I got _somewhat_ stable perf numbers for compiling BasicWebApi:

Baseline: 10.451 s, 10.542 s, 10.513 s, 10.330 s, 10.292 s
PR: 10.131 s, 10.262 s, 10.242 s, 10.276 s, 10.287 s
`Dictionary<(string, string), TypeDefinitionHandle>`: 10.365 s, 10.309 s, 10.200 s, 10.364 s, 10.245 s

And this is for the PR version that hashes things a lot worse than what we can do if it's done properly (not byte-by-byte).

We can go with dictionary, but it's going to materialize more strings than we typically need. Right now this is amortized by NameMangler also materializing all the strings in the assembly, but that's something on my todo list to fix at some point (i.e. it's only going to get worse).

Maybe go with the dictionary now but leave a TODO that this is leaving perf on the table?

> The only thing I would be somewhat concerned about is the initial perf cost of building the hashtable, that might pop up as startup slowdown

Yeah, there's a risk that if there's an assembly with lots of types and we only need a small handful, we might not be able to pay for the hashtable building. On average the type we're looking for is going to be in the middle, so on average we can only save half of the work, even if only one type was accessed. But this would add more synchronization concerns - this needs to be threadsafe. And at the point we'd add locking it might end up being a wash.