@MichalStrehovsky 2023-04-05
> Also, would it be better to use the split IBC data per assembly, like what we do for R2R?

Can you send me a pointer? I don't know what this refers to.