@jkotas 2023-04-05
> Can you send me a pointer? I don't know what this refers to.

`--embed-pgo-data` command line option for crossgen2: https://github.com/dotnet/runtime/blob/3c74d65445f20e09a9542940a1768c5de6bfe1f9/src/coreclr/tools/aot/crossgen2/Crossgen2RootCommand.cs#L58 . Implemented by https://github.com/dotnet/runtime/blob/main/src/coreclr/tools/aot/ILCompiler.ReadyToRun/Compiler/DependencyAnalysis/ReadyToRun/InstrumentationDataTableNode.cs . 

It indexes the profile data to allows us to quickly find the profile data for given method without reading everything.