@jkotas 2023-04-05
> Maybe go with the dictionary now but leave a TODO that this is leaving perf on the table?

Sounds good to me.