@MichalStrehovsky 2023-04-05
> It indexes the profile data to allows us to quickly find the profile data for given method without reading everything.

We'll be able to read this information when #67080 is done - but currently the framework ILC sees is not R2R compiled.