@jkotas 2023-04-05
> We'll be able to read this information when https://github.com/dotnet/runtime/issues/67080 is done - but currently the framework ILC sees is not R2R compiled.

The most impactful profile data are attached to CoreLib that won't be automatically covered by this.